backend/requirements.txt
fastapi==0.115.12
uvicorn==0.34.3
psycopg==3.2.3
pgvector==0.4.1
python-dotenv==1.1.0
pandas==2.3.0
numpy==2.3.0
openai==1.86.0
pydantic==2.11.5
pydantic-settings==2.2.1
requests==2.32.4
python-dateutil==2.9.0.post0
pytz==2025.2
psutil==6.1.1